+++ common/src/main/java/com/alien/common/registry/init/item/AlienChitinBlockItems.java
@@ -1,0 +1,146 @@
+package com.alien.common.registry.init.item;
+
+import com.alien.common.registry.init.block.AlienChitinBlocks;
+import com.avp.common.registry.AVPDeferredHolder;
+import com.avp.common.registry.init.item.AVPBlockItems;
+import net.minecraft.world.item.BlockItem;
+
+public class AlienChitinBlockItems {
+
+    public static final AVPDeferredHolder<BlockItem> ABERRANT_CHITIN_BLOCK = AVPBlockItems.register(
+        "aberrant_chitin_block",
+        AlienChitinBlocks.ABERRANT_CHITIN_BLOCK
+    );
+
+    public static final AVPDeferredHolder<BlockItem> ABERRANT_CHITIN_BLOCK_SLAB = AVPBlockItems.register(
+        "aberrant_chitin_block_slab",
+        AlienChitinBlocks.ABERRANT_CHITIN_BLOCK_SLAB
+    );
+
+    public static final AVPDeferredHolder<BlockItem> ABERRANT_CHITIN_BLOCK_STAIRS = AVPBlockItems.register(
+        "aberrant_chitin_block_stairs",
+        AlienChitinBlocks.ABERRANT_CHITIN_BLOCK_STAIRS
+    );
+
+    public static final AVPDeferredHolder<BlockItem> ABERRANT_CHITIN_BRICKS = AVPBlockItems.register(
+        "aberrant_chitin_bricks",
+        AlienChitinBlocks.ABERRANT_CHITIN_BRICKS
+    );
+
+    public static final AVPDeferredHolder<BlockItem> ABERRANT_CHITIN_BRICK_SLAB = AVPBlockItems.register(
+        "aberrant_chitin_brick_slab",
+        AlienChitinBlocks.ABERRANT_CHITIN_BRICK_SLAB
+    );
+
+    public static final AVPDeferredHolder<BlockItem> ABERRANT_CHITIN_BRICK_STAIRS = AVPBlockItems.register(
+        "aberrant_chitin_brick_stairs",
+        AlienChitinBlocks.ABERRANT_CHITIN_BRICK_STAIRS
+    );
+
+    public static final AVPDeferredHolder<BlockItem> CHITIN_BLOCK = AVPBlockItems.register(
+        "chitin_block",
+        AlienChitinBlocks.CHITIN_BLOCK
+    );
+
+    public static final AVPDeferredHolder<BlockItem> CHITIN_BLOCK_SLAB = AVPBlockItems.register(
+        "chitin_block_slab",
+        AlienChitinBlocks.CHITIN_BLOCK_SLAB
+    );
+
+    public static final AVPDeferredHolder<BlockItem> CHITIN_BLOCK_STAIRS = AVPBlockItems.register(
+        "chitin_block_stairs",
+        AlienChitinBlocks.CHITIN_BLOCK_STAIRS
+    );
+
+    public static final AVPDeferredHolder<BlockItem> CHITIN_BRICKS = AVPBlockItems.register(
+        "chitin_bricks",
+        AlienChitinBlocks.CHITIN_BRICKS
+    );
+
+    public static final AVPDeferredHolder<BlockItem> CHITIN_BRICK_SLAB = AVPBlockItems.register(
+        "chitin_brick_slab",
+        AlienChitinBlocks.CHITIN_BRICK_SLAB
+    );
+
+    public static final AVPDeferredHolder<BlockItem> CHITIN_BRICK_STAIRS = AVPBlockItems.register(
+        "chitin_brick_stairs",
+        AlienChitinBlocks.CHITIN_BRICK_STAIRS
+    );
+
+    public static final AVPDeferredHolder<BlockItem> NETHER_CHITIN_BLOCK = AVPBlockItems.register(
+        "nether_chitin_block",
+        AlienChitinBlocks.NETHER_CHITIN_BLOCK
+    );
+
+    public static final AVPDeferredHolder<BlockItem> NETHER_CHITIN_BLOCK_SLAB = AVPBlockItems.register(
+        "nether_chitin_block_slab",
+        AlienChitinBlocks.NETHER_CHITIN_BLOCK_SLAB
+    );
+
+    public static final AVPDeferredHolder<BlockItem> NETHER_CHITIN_BLOCK_STAIRS = AVPBlockItems.register(
+        "nether_chitin_block_stairs",
+        AlienChitinBlocks.NETHER_CHITIN_BLOCK_STAIRS
+    );
+
+    public static final AVPDeferredHolder<BlockItem> NETHER_CHITIN_BRICKS = AVPBlockItems.register(
+        "nether_chitin_bricks",
+        AlienChitinBlocks.NETHER_CHITIN_BRICKS
+    );
+
+    public static final AVPDeferredHolder<BlockItem> NETHER_CHITIN_BRICK_SLAB = AVPBlockItems.register(
+        "nether_chitin_brick_slab",
+        AlienChitinBlocks.NETHER_CHITIN_BRICK_SLAB
+    );
+
+    public static final AVPDeferredHolder<BlockItem> NETHER_CHITIN_BRICK_STAIRS = AVPBlockItems.register(
+        "nether_chitin_brick_stairs",
+        AlienChitinBlocks.NETHER_CHITIN_BRICK_STAIRS
+    );
+
+    public static final AVPDeferredHolder<BlockItem> POLISHED_ABERRANT_CHITIN = AVPBlockItems.register(
+        "polished_aberrant_chitin",
+        AlienChitinBlocks.POLISHED_ABERRANT_CHITIN
+    );
+
+    public static final AVPDeferredHolder<BlockItem> POLISHED_ABERRANT_CHITIN_SLAB = AVPBlockItems.register(
+        "polished_aberrant_chitin_slab",
+        AlienChitinBlocks.POLISHED_ABERRANT_CHITIN_SLAB
+    );
+
+    public static final AVPDeferredHolder<BlockItem> POLISHED_ABERRANT_CHITIN_STAIRS = AVPBlockItems.register(
+        "polished_aberrant_chitin_stairs",
+        AlienChitinBlocks.POLISHED_ABERRANT_CHITIN_STAIRS
+    );
+
+    public static final AVPDeferredHolder<BlockItem> POLISHED_CHITIN = AVPBlockItems.register(
+        "polished_chitin",
+        AlienChitinBlocks.POLISHED_CHITIN
+    );
+
+    public static final AVPDeferredHolder<BlockItem> POLISHED_CHITIN_SLAB = AVPBlockItems.register(
+        "polished_chitin_slab",
+        AlienChitinBlocks.POLISHED_CHITIN_SLAB
+    );
+
+    public static final AVPDeferredHolder<BlockItem> POLISHED_CHITIN_STAIRS = AVPBlockItems.register(
+        "polished_chitin_stairs",
+        AlienChitinBlocks.POLISHED_CHITIN_STAIRS
+    );
+
+    public static final AVPDeferredHolder<BlockItem> POLISHED_NETHER_CHITIN = AVPBlockItems.register(
+        "polished_nether_chitin",
+        AlienChitinBlocks.POLISHED_NETHER_CHITIN
+    );
+
+    public static final AVPDeferredHolder<BlockItem> POLISHED_NETHER_CHITIN_SLAB = AVPBlockItems.register(
+        "polished_nether_chitin_slab",
+        AlienChitinBlocks.POLISHED_NETHER_CHITIN_SLAB
+    );
+
+    public static final AVPDeferredHolder<BlockItem> POLISHED_NETHER_CHITIN_STAIRS = AVPBlockItems.register(
+        "polished_nether_chitin_stairs",
+        AlienChitinBlocks.POLISHED_NETHER_CHITIN_STAIRS
+    );
+
+    public static void initialize() {}
+}